refactor: plugin 서비스 계층 추출 — router 1,616줄 → 301줄
- plugin/service.py 신설: 전 엔드포인트의 비즈니스 로직을 키워드 전용
  인자 서비스 함수로 복사-이동 (응답 형태·상태코드·에러 메시지 불변)
- upload_plugin(250줄+)을 단계 함수로 분해:
  _backup_existing_plugin → _save_plugin_files → _upsert_plugin
  (실패 시 _rollback_upload로 파일/백업 원복)
- Docker 빌드 위임부를 plugin/builder.py로 분리 (utils.py 내부는 PR-9)
- router는 파싱/deps/service 호출/응답 반환만 담당
- tests/unit/test_plugin_service.py 신규 14건 (롤백 실패 주입 포함,
  Docker·파일시스템 모킹)
- characterization patch 경로만 서비스 모듈로 갱신 (기대값 불변)

과도기 규칙: 서비스의 HTTPException은 PR-8에서 도메인 예외로 일괄 전환
